+++ tests/test_ca_cert.py
@@ -13,7 +13,7 @@ from client import Sendonly
 # https://letsencrypt.org/certificates/
 # ISRG Root X1
 
-letsencrypt_org_ca_cert = """
+letsencrypt_org_ca_cert = b"""
 -----BEGIN CERTIFICATE-----
 MIIFazCCA1OgAwIBAgIRAIIQz7DSQONZRGPgu2OCiwAwDQYJKoZIhvcNAQELBQAw
 TzELMAkGA1UEBhMCVVMxKTAnBgNVBAoTIEludGVybmV0IFNlY3VyaXR5IFJlc2Vh
@@ -49,7 +49,7 @@ emyPxgcYxn/eR44/KJ4EBs+lVDR3veyJm+kXQ99b21/+jh5Xos1AnX5iItreGCc=
 
 # https://pki.goog/repository/
 
-pki_goog_ca_cert = """
+pki_goog_ca_cert = b"""
 -----BEGIN CERTIFICATE-----
 MIICCTCCAY6gAwIBAgINAgPlwGjvYxqccpBQUjAKBggqhkjOPQQDAzBHMQswCQYD
 VQQGEwJVUzEiMCAGA1UEChMZR29vZ2xlIFRydXN0IFNlcnZpY2VzIExMQzEUMBIG