+++ CHANGES.md
@@ -14,6 +14,8 @@
 - [CHANGE] client_cert と client_key の指定にはパスではなく中身の文字列を指定するようにする
   - C++ SDK 側の仕様変更に追従する
   - @voluntas
+- [CHANGE] ca_cert, client_cert, client_key の指定には str ではなく bytes を使うようにする
+  - @tnoho
 - [ADD] サーバー証明書チェック用の CA 証明書を指定できるようにする
   - `Sora.create_connection()` の引数に `ca_cert` を追加する
   - @voluntas